New version of my rc.lua
I now use the cal lua module instead of calendar2. I modified the module
itself a bit to be cleaner and more compact. More changes will come to
enable three months diplay similar to the "cal -3" system command.

Also added a bunch more widgets to the wibox and cleaned up the file
quite a bit as well.

Made the test script for awesome configs more robust.

Added a color for the labels to the theme file.