Added new features to my awesome WM setup.
New widgets on the desktop include a power control menu which can be
used to log off, turn the computer off, or reboot. There's also a
tooltip to display network information when hovering over the speed
widget. Similarly, there's a new tooltip displaying system information
when hovering over the renamed Sys label.

Changed the formatting of the calendar widget to match the formatting
used in the other tooltips.

Added the iconds that I use on my desktop. They are actually only used
in the power control menu.