Minor usability improvements.
Turns out that some of my co-workers like to use the .hpp extension for
C++ header files. So I added that to my 'fc' and 'fh' aliases.

The '%%' expands now to a relative path instead of the file's full path
if the file is opened with a relative path.

The install script now only warns of missing binaries instead of erroring
and exiting.