tools: Add LUKS wrapper for easy encrypted folders.
You can use it like so:

1) sudo crypt-helper.py create --size 20M foo.img

This will create a 20 MiB file called foo.img. Its contents will be a
LUKS-encrypted block device that contains an EXT4 partition. This step
will ask you to select a password to assign to the encrypted volume.

2) sudo crypt-helper.py mount foo.img /mnt/

Mounts the EXT4 file system inside of foo.img to /mnt/. This will ask
for the volume's password you specified in step 1.

3) sudo crypt-helper.py unmount foo.img /mnt/

Unmount the EXT4 partition and close the LUKS mapper so that it has to
be unlocked again.