bash: Parse ~ and variables in ~/.bash/paths.
Before this change the paths specified in ~/.bash/paths were getting
prepended to PATH literally. So all the ~ and other bash variables
were not getting interpreted.

Bash itself was perfectly happy with this as it could simply evaluate
those characters every time I tried to tab-complete. Other programs,
however, are unable to parse it properly. This came up when I was
trying to the autopep8 vim plugin to work. It just wasn't picking up
the path ~/.local/bin because there was no ~ vim in vim's opinion.

I changed it so that now the paths in those files are expanded fully
and all programs should be happy with the result.