vim: Replace a.vim with vim-fswitch
a.vim is quite outdated I realized and hard-coded a bunch of shortcuts
that interfered with what I've already configured for myself. The
worst offender was "<Leader>is" in insert mode. Since I have <Leader>
mapped to Space it would trigger every time I typed " is" as text. I
just can't deal with that.

I still have to submit a patch to vim-fswitch to deal with .cc files,
but I'll get to that.