vim: Make formatoptions a little more useful
This enables 'j' and 'l' in formatoptions. From the help:

l	Long lines are not broken in insert mode: When a line was
        longer than 'textwidth' when the insert command started, Vim
        does not automatically format it.
j	Where it makes sense, remove a comment leader when joining
        lines.  For example, joining:
		int i;   // the index ~
		         // in the list ~
	Becomes:
		int i;   // the index in the list ~

I'm also enabling 'n' for Markdown and plain text files for
convenience. Also turn off '2' at the same time as per the help:

n	When formatting text, recognize numbered lists.  This actually
        uses the 'formatlistpat' option, thus any kind of list can be
        used.  The indent of the text after the number is used for the
        next line.  The default is to find a number, optionally
        followed by '.', ':', ')', ']' or '}'.  Note that 'autoindent'
        must be set too.  Doesn't work well together with "2".
	Example: >
		1. the first item
		   wraps
		2. the second item