awesome: Add some primitive battery widgets for 4.0
I really needed this. It's very ugly, but it'll work for now.

To make this work I actually had to mark a bunch of variables in the
battery widget as local. The "remaining" variable in the widget
interfered with the "remaining" variable I used in my rc file.

Also removed the ability to click on the layout box. I kept doing that
accidentally which is really annoying.