tmux: Disable the tmux-copycat plugin for now
There are some weird issues with the shortcuts that it registers.
In particular, it eats the n/N keys I type and delays them a
non-trivial amount of time.

For example, if I type "Nowhere" fast enough, it might show up as
"owNhere" in the actual terminal. This gets worse on slower
connections over SSH.