vim: Fix slow pastes when authoring git commits
The syntax highlighting makes pastes take a really long time.
Sometimes even a few minutes. The easy solution here is to turn off
syntax highlighting while in insert mode. It's not ideal, but it does
the job for now. It would be cool if I could figure out how to only
disable syntax highlighting when in paste mode.