tmux: Make moving windows around more pleasant
Before this patch, the windows would swap as expected, but you would
"switch" to the window that was swapped in. It seems that tmux tracks
the currently selected window by index. So in order to make it more
user friendly, we also need to switch to the tab we just swapped.

This makes it a lot easier to move tabs around.